add exe-1 parenthesis
12:42 to 12:54

I started this one by making the test cases. I drew the trees for `(1+2)*3`, `1+2*3`, `(1+2)*(3+4)` and `(1+2)*3+4` and studied them for a bit.

First I thought it had to do with how deep the left subtree was, but the `(1+2)*(3+4)` case busted that idea.

Then I thought about the concept of priority, and how there were rules for priority. Some operations had the same priority, while others had higher or lower priority.

With this in mind and looking at the trees for `(1+2)*(3+4)` and `(1+2)*3+4`, it occurred to me that parenthesis were needed when the `toString()` caller had higher priority than the callee.

This led me to think each node should have some priority logic. But if I wanted to extend the set with more operations, they couldn't check operation identify. Instead I opted with a number system. I only really allowed for values(prio 0), sum/subtraction (prio 1), and multiplication/division (prio 2). I know other operations like exponentiation have higher priority but would need to go consult a table for the full list. This choise would have to be revisited with better domain knowledge.

I started modifying the Multiplication `toString` to check the priorities of its values and wrapping each in parenthesis if needed.

Then I thought it would be better to do it the other way around: wrap everything in parenthesis based on the caller prio.

So I modified the Node interface to send the priority, and added a helper `wrapParens` helper function since all operations would use the same code. I also thought of making `Node` a class and adding `wrapParens` as a method but it didn't seem worth to bother with it right now, it's just a refactor after testing things work as intended.

At the end I also thought the tests would have been easier to write if I had declared individual variables for `Value(1)`, 2 and 3. But at this point it didn't matter.