add tentative exe-1 division
Add tentative solution for https://github.com/filipesilva/misko-exes/commit/b1d7c08f5c93d3fa24043d854ab4950ff5d12132#commitcomment-27539316.

This problem made me think for a while.

I came to the conclusion that the current implementation based on passing the caller priority to child nodes expected that order between the two operands did not matter (commutative property).

This is true for sums (1+2 == 2+1) and multiplications (1*2 == 2*1) but it is not true for subtractions (1-2 != 2-1) or divisions (1/2 != 2/1).

To express this using the priority system, I added subtration and division as always having infinite priority for the purposes of determining whether the second operand should be wrapped.

This causes any non-values to always have parenthesis when in the second operand for these two operations, while retaining the existing priority for the first operand.

At this point I am not so sure this is the best solution. It solves the specific cases I tested but I would feel a lot better by looking up the theory behind arithmetic, and exploring how some more exiting operators would behave.